Moved ASTNode away from workspace_context
