+++ aderyn_core/src/context/mod.rs
@@ -1,4 +1,5 @@
 pub mod browser;
 pub mod capturable;
 pub mod macros;
+pub mod meta_workspace;
 pub mod workspace_context;
