+++ aderyn_core/src/context/workspace_context.rs
@@ -6,437 +6,8 @@ use std::collections::HashMap;
 
 use super::browser::GetImmediateParent;
 use super::capturable::Capturable;
-use super::macros::{
-    generate_ast_methods, generate_visit_methods_for_workspace_context_with_insert_node,
-};
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum ASTNode {
-    ArrayTypeName(ArrayTypeName),
-    Assignment(Assignment),
-    BinaryOperation(BinaryOperation),
-    Block(Block),
-    Conditional(Conditional),
-    ContractDefinition(ContractDefinition),
-    ElementaryTypeName(ElementaryTypeName),
-    ElementaryTypeNameExpression(ElementaryTypeNameExpression),
-    EmitStatement(EmitStatement),
-    EnumDefinition(EnumDefinition),
-    EnumValue(EnumValue),
-    EventDefinition(EventDefinition),
-    ErrorDefinition(ErrorDefinition),
-    ExpressionStatement(ExpressionStatement),
-    FunctionCall(FunctionCall),
-    FunctionCallOptions(FunctionCallOptions),
-    FunctionDefinition(FunctionDefinition),
-    FunctionTypeName(FunctionTypeName),
-    ForStatement(ForStatement),
-    Identifier(Identifier),
-    IdentifierPath(IdentifierPath),
-    IfStatement(IfStatement),
-    ImportDirective(ImportDirective),
-    IndexAccess(IndexAccess),
-    IndexRangeAccess(IndexRangeAccess),
-    InheritanceSpecifier(InheritanceSpecifier),
-    InlineAssembly(InlineAssembly),
-    Literal(Literal),
-    MemberAccess(MemberAccess),
-    NewExpression(NewExpression),
-    Mapping(Mapping),
-    ModifierDefinition(ModifierDefinition),
-    ModifierInvocation(ModifierInvocation),
-    OverrideSpecifier(OverrideSpecifier),
-    ParameterList(ParameterList),
-    PragmaDirective(PragmaDirective),
-    Return(Return),
-    RevertStatement(RevertStatement),
-    SourceUnit(SourceUnit),
-    StructDefinition(StructDefinition),
-    StructuredDocumentation(StructuredDocumentation),
-    TryStatement(TryStatement),
-    TryCatchClause(TryCatchClause),
-    TupleExpression(TupleExpression),
-    UnaryOperation(UnaryOperation),
-    UserDefinedTypeName(UserDefinedTypeName),
-    UserDefinedValueTypeDefinition(UserDefinedValueTypeDefinition),
-    UsingForDirective(UsingForDirective),
-    VariableDeclaration(VariableDeclaration),
-    VariableDeclarationStatement(VariableDeclarationStatement),
-    WhileStatement(WhileStatement),
-    DoWhileStatement(DoWhileStatement),
-    Break(Break),
-    Continue(Continue),
-    PlaceholderStatement(PlaceholderStatement),
-}
-
-impl ASTNode {
-    pub fn node_type(&self) -> NodeType {
-        match self {
-            ASTNode::ArrayTypeName(_) => NodeType::ArrayTypeName,
-            ASTNode::Assignment(_) => NodeType::Assignment,
-            ASTNode::BinaryOperation(_) => NodeType::BinaryOperation,
-            ASTNode::Block(_) => NodeType::Block,
-            ASTNode::Conditional(_) => NodeType::Conditional,
-            ASTNode::ContractDefinition(_) => NodeType::ContractDefinition,
-            ASTNode::ElementaryTypeName(_) => NodeType::ElementaryTypeName,
-            ASTNode::ElementaryTypeNameExpression(_) => NodeType::ElementaryTypeNameExpression,
-            ASTNode::EmitStatement(_) => NodeType::EmitStatement,
-            ASTNode::EnumDefinition(_) => NodeType::EnumDefinition,
-            ASTNode::EnumValue(_) => NodeType::EnumValue,
-            ASTNode::EventDefinition(_) => NodeType::EventDefinition,
-            ASTNode::ErrorDefinition(_) => NodeType::ErrorDefinition,
-            ASTNode::ExpressionStatement(_) => NodeType::ExpressionStatement,
-            ASTNode::FunctionCall(_) => NodeType::FunctionCall,
-            ASTNode::FunctionCallOptions(_) => NodeType::FunctionCallOptions,
-            ASTNode::FunctionDefinition(_) => NodeType::FunctionDefinition,
-            ASTNode::FunctionTypeName(_) => NodeType::FunctionTypeName,
-            ASTNode::ForStatement(_) => NodeType::ForStatement,
-            ASTNode::Identifier(_) => NodeType::Identifier,
-            ASTNode::IdentifierPath(_) => NodeType::IdentifierPath,
-            ASTNode::IfStatement(_) => NodeType::IfStatement,
-            ASTNode::ImportDirective(_) => NodeType::ImportDirective,
-            ASTNode::IndexAccess(_) => NodeType::IndexAccess,
-            ASTNode::IndexRangeAccess(_) => NodeType::IndexRangeAccess,
-            ASTNode::InheritanceSpecifier(_) => NodeType::InheritanceSpecifier,
-            ASTNode::InlineAssembly(_) => NodeType::InlineAssembly,
-            ASTNode::Literal(_) => NodeType::Literal,
-            ASTNode::MemberAccess(_) => NodeType::MemberAccess,
-            ASTNode::NewExpression(_) => NodeType::NewExpression,
-            ASTNode::Mapping(_) => NodeType::Mapping,
-            ASTNode::ModifierDefinition(_) => NodeType::ModifierDefinition,
-            ASTNode::ModifierInvocation(_) => NodeType::ModifierInvocation,
-            ASTNode::OverrideSpecifier(_) => NodeType::OverrideSpecifier,
-            ASTNode::ParameterList(_) => NodeType::ParameterList,
-            ASTNode::PragmaDirective(_) => NodeType::PragmaDirective,
-            ASTNode::Return(_) => NodeType::Return,
-            ASTNode::RevertStatement(_) => NodeType::RevertStatement,
-            ASTNode::SourceUnit(_) => NodeType::SourceUnit,
-            ASTNode::StructDefinition(_) => NodeType::StructDefinition,
-            ASTNode::StructuredDocumentation(_) => NodeType::StructuredDocumentation,
-            ASTNode::TryStatement(_) => NodeType::TryStatement,
-            ASTNode::TryCatchClause(_) => NodeType::TryCatchClause,
-            ASTNode::TupleExpression(_) => NodeType::TupleExpression,
-            ASTNode::UnaryOperation(_) => NodeType::UnaryOperation,
-            ASTNode::UserDefinedTypeName(_) => NodeType::UserDefinedTypeName,
-            ASTNode::UserDefinedValueTypeDefinition(_) => NodeType::UserDefinedValueTypeDefinition,
-            ASTNode::UsingForDirective(_) => NodeType::UsingForDirective,
-            ASTNode::VariableDeclaration(_) => NodeType::VariableDeclaration,
-            ASTNode::VariableDeclarationStatement(_) => NodeType::VariableDeclarationStatement,
-            ASTNode::WhileStatement(_) => NodeType::WhileStatement,
-            ASTNode::DoWhileStatement(_) => NodeType::DoWhileStatement,
-            ASTNode::Break(_) => NodeType::Break,
-            ASTNode::Continue(_) => NodeType::Continue,
-            ASTNode::PlaceholderStatement(_) => NodeType::PlaceholderStatement,
-        }
-    }
-
-    pub fn id(&self) -> Option<NodeID> {
-        match self {
-            ASTNode::ArrayTypeName(n) => Some(n.id),
-            ASTNode::Assignment(n) => Some(n.id),
-            ASTNode::BinaryOperation(n) => Some(n.id),
-            ASTNode::Block(n) => Some(n.id),
-            ASTNode::Conditional(n) => Some(n.id),
-            ASTNode::ContractDefinition(n) => Some(n.id),
-            ASTNode::ElementaryTypeName(n) => Some(n.id),
-            ASTNode::ElementaryTypeNameExpression(n) => Some(n.id),
-            ASTNode::EmitStatement(n) => Some(n.id),
-            ASTNode::EnumDefinition(n) => Some(n.id),
-            ASTNode::EnumValue(n) => Some(n.id),
-            ASTNode::EventDefinition(n) => Some(n.id),
-            ASTNode::ErrorDefinition(n) => Some(n.id),
-            ASTNode::ExpressionStatement(n) => Some(n.id),
-            ASTNode::FunctionCall(n) => Some(n.id),
-            ASTNode::FunctionCallOptions(n) => Some(n.id),
-            ASTNode::FunctionDefinition(n) => Some(n.id),
-            ASTNode::FunctionTypeName(n) => Some(n.id),
-            ASTNode::ForStatement(n) => Some(n.id),
-            ASTNode::Identifier(n) => Some(n.id),
-            ASTNode::IdentifierPath(n) => Some(n.id),
-            ASTNode::IfStatement(n) => Some(n.id),
-            ASTNode::ImportDirective(n) => Some(n.id),
-            ASTNode::IndexAccess(n) => Some(n.id),
-            ASTNode::IndexRangeAccess(n) => Some(n.id),
-            ASTNode::InheritanceSpecifier(n) => Some(n.id),
-            ASTNode::InlineAssembly(n) => Some(n.id),
-            ASTNode::Literal(n) => Some(n.id),
-            ASTNode::MemberAccess(n) => Some(n.id),
-            ASTNode::NewExpression(n) => Some(n.id),
-            ASTNode::Mapping(n) => Some(n.id),
-            ASTNode::ModifierDefinition(n) => Some(n.id),
-            ASTNode::ModifierInvocation(n) => Some(n.id),
-            ASTNode::OverrideSpecifier(n) => Some(n.id),
-            ASTNode::ParameterList(n) => Some(n.id),
-            ASTNode::PragmaDirective(n) => Some(n.id),
-            ASTNode::Return(n) => Some(n.id),
-            ASTNode::RevertStatement(n) => Some(n.id),
-            ASTNode::SourceUnit(n) => Some(n.id),
-            ASTNode::StructDefinition(n) => Some(n.id),
-            ASTNode::StructuredDocumentation(n) => Some(n.id),
-            ASTNode::TryStatement(n) => Some(n.id),
-            ASTNode::TryCatchClause(n) => Some(n.id),
-            ASTNode::TupleExpression(n) => Some(n.id),
-            ASTNode::UnaryOperation(n) => Some(n.id),
-            ASTNode::UserDefinedTypeName(n) => Some(n.id),
-            ASTNode::UserDefinedValueTypeDefinition(n) => Some(n.id),
-            ASTNode::UsingForDirective(n) => Some(n.id),
-            ASTNode::VariableDeclaration(n) => Some(n.id),
-            ASTNode::VariableDeclarationStatement(n) => Some(n.id),
-            ASTNode::WhileStatement(n) => Some(n.id),
-            ASTNode::DoWhileStatement(n) => Some(n.id),
-            ASTNode::Break(n) => Some(n.id),
-            ASTNode::Continue(n) => Some(n.id),
-            ASTNode::PlaceholderStatement(n) => Some(n.id),
-        }
-    }
-}
-
-impl Node for ASTNode {
-    fn accept(&self, visitor: &mut impl ASTConstVisitor) -> eyre::Result<()> {
-        match self {
-            ASTNode::ArrayTypeName(n) => n.accept(visitor),
-            ASTNode::Assignment(n) => n.accept(visitor),
-            ASTNode::BinaryOperation(n) => n.accept(visitor),
-            ASTNode::Block(n) => n.accept(visitor),
-            ASTNode::Conditional(n) => n.accept(visitor),
-            ASTNode::ContractDefinition(n) => n.accept(visitor),
-            ASTNode::ElementaryTypeName(n) => n.accept(visitor),
-            ASTNode::ElementaryTypeNameExpression(n) => n.accept(visitor),
-            ASTNode::EmitStatement(n) => n.accept(visitor),
-            ASTNode::EnumDefinition(n) => n.accept(visitor),
-            ASTNode::EnumValue(n) => n.accept(visitor),
-            ASTNode::EventDefinition(n) => n.accept(visitor),
-            ASTNode::ErrorDefinition(n) => n.accept(visitor),
-            ASTNode::ExpressionStatement(n) => n.accept(visitor),
-            ASTNode::FunctionCall(n) => n.accept(visitor),
-            ASTNode::FunctionCallOptions(n) => n.accept(visitor),
-            ASTNode::FunctionDefinition(n) => n.accept(visitor),
-            ASTNode::FunctionTypeName(n) => n.accept(visitor),
-            ASTNode::ForStatement(n) => n.accept(visitor),
-            ASTNode::Identifier(n) => n.accept(visitor),
-            ASTNode::IdentifierPath(n) => n.accept(visitor),
-            ASTNode::IfStatement(n) => n.accept(visitor),
-            ASTNode::ImportDirective(n) => n.accept(visitor),
-            ASTNode::IndexAccess(n) => n.accept(visitor),
-            ASTNode::IndexRangeAccess(n) => n.accept(visitor),
-            ASTNode::InheritanceSpecifier(n) => n.accept(visitor),
-            ASTNode::InlineAssembly(n) => n.accept(visitor),
-            ASTNode::Literal(n) => n.accept(visitor),
-            ASTNode::MemberAccess(n) => n.accept(visitor),
-            ASTNode::NewExpression(n) => n.accept(visitor),
-            ASTNode::Mapping(n) => n.accept(visitor),
-            ASTNode::ModifierDefinition(n) => n.accept(visitor),
-            ASTNode::ModifierInvocation(n) => n.accept(visitor),
-            ASTNode::OverrideSpecifier(n) => n.accept(visitor),
-            ASTNode::ParameterList(n) => n.accept(visitor),
-            ASTNode::PragmaDirective(n) => n.accept(visitor),
-            ASTNode::Return(n) => n.accept(visitor),
-            ASTNode::RevertStatement(n) => n.accept(visitor),
-            ASTNode::SourceUnit(n) => n.accept(visitor),
-            ASTNode::StructDefinition(n) => n.accept(visitor),
-            ASTNode::StructuredDocumentation(n) => n.accept(visitor),
-            ASTNode::TryStatement(n) => n.accept(visitor),
-            ASTNode::TryCatchClause(n) => n.accept(visitor),
-            ASTNode::TupleExpression(n) => n.accept(visitor),
-            ASTNode::UnaryOperation(n) => n.accept(visitor),
-            ASTNode::UserDefinedTypeName(n) => n.accept(visitor),
-            ASTNode::UserDefinedValueTypeDefinition(n) => n.accept(visitor),
-            ASTNode::UsingForDirective(n) => n.accept(visitor),
-            ASTNode::VariableDeclaration(n) => n.accept(visitor),
-            ASTNode::VariableDeclarationStatement(n) => n.accept(visitor),
-            ASTNode::WhileStatement(n) => n.accept(visitor),
-            ASTNode::DoWhileStatement(n) => n.accept(visitor),
-            ASTNode::Break(n) => n.accept(visitor),
-            ASTNode::Continue(n) => n.accept(visitor),
-            ASTNode::PlaceholderStatement(n) => n.accept(visitor),
-        }
-    }
-
-    fn accept_metadata(&self, visitor: &mut impl ASTConstVisitor) -> eyre::Result<()> {
-        match self {
-            ASTNode::ArrayTypeName(n) => n.accept_metadata(visitor),
-            ASTNode::Assignment(n) => n.accept_metadata(visitor),
-            ASTNode::BinaryOperation(n) => n.accept_metadata(visitor),
-            ASTNode::Block(n) => n.accept_metadata(visitor),
-            ASTNode::Conditional(n) => n.accept_metadata(visitor),
-            ASTNode::ContractDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::ElementaryTypeName(n) => n.accept_metadata(visitor),
-            ASTNode::ElementaryTypeNameExpression(n) => n.accept_metadata(visitor),
-            ASTNode::EmitStatement(n) => n.accept_metadata(visitor),
-            ASTNode::EnumDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::EnumValue(n) => n.accept_metadata(visitor),
-            ASTNode::EventDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::ErrorDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::ExpressionStatement(n) => n.accept_metadata(visitor),
-            ASTNode::FunctionCall(n) => n.accept_metadata(visitor),
-            ASTNode::FunctionCallOptions(n) => n.accept_metadata(visitor),
-            ASTNode::FunctionDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::FunctionTypeName(n) => n.accept_metadata(visitor),
-            ASTNode::ForStatement(n) => n.accept_metadata(visitor),
-            ASTNode::Identifier(n) => n.accept_metadata(visitor),
-            ASTNode::IdentifierPath(n) => n.accept_metadata(visitor),
-            ASTNode::IfStatement(n) => n.accept_metadata(visitor),
-            ASTNode::ImportDirective(n) => n.accept_metadata(visitor),
-            ASTNode::IndexAccess(n) => n.accept_metadata(visitor),
-            ASTNode::IndexRangeAccess(n) => n.accept_metadata(visitor),
-            ASTNode::InheritanceSpecifier(n) => n.accept_metadata(visitor),
-            ASTNode::InlineAssembly(n) => n.accept_metadata(visitor),
-            ASTNode::Literal(n) => n.accept_metadata(visitor),
-            ASTNode::MemberAccess(n) => n.accept_metadata(visitor),
-            ASTNode::NewExpression(n) => n.accept_metadata(visitor),
-            ASTNode::Mapping(n) => n.accept_metadata(visitor),
-            ASTNode::ModifierDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::ModifierInvocation(n) => n.accept_metadata(visitor),
-            ASTNode::OverrideSpecifier(n) => n.accept_metadata(visitor),
-            ASTNode::ParameterList(n) => n.accept_metadata(visitor),
-            ASTNode::PragmaDirective(n) => n.accept_metadata(visitor),
-            ASTNode::Return(n) => n.accept_metadata(visitor),
-            ASTNode::RevertStatement(n) => n.accept_metadata(visitor),
-            ASTNode::SourceUnit(n) => n.accept_metadata(visitor),
-            ASTNode::StructDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::StructuredDocumentation(n) => n.accept_metadata(visitor),
-            ASTNode::TryStatement(n) => n.accept_metadata(visitor),
-            ASTNode::TryCatchClause(n) => n.accept_metadata(visitor),
-            ASTNode::TupleExpression(n) => n.accept_metadata(visitor),
-            ASTNode::UnaryOperation(n) => n.accept_metadata(visitor),
-            ASTNode::UserDefinedTypeName(n) => n.accept_metadata(visitor),
-            ASTNode::UserDefinedValueTypeDefinition(n) => n.accept_metadata(visitor),
-            ASTNode::UsingForDirective(n) => n.accept_metadata(visitor),
-            ASTNode::VariableDeclaration(n) => n.accept_metadata(visitor),
-            ASTNode::VariableDeclarationStatement(n) => n.accept_metadata(visitor),
-            ASTNode::WhileStatement(n) => n.accept_metadata(visitor),
-            ASTNode::DoWhileStatement(n) => n.accept_metadata(visitor),
-            ASTNode::Break(n) => n.accept_metadata(visitor),
-            ASTNode::Continue(n) => n.accept_metadata(visitor),
-            ASTNode::PlaceholderStatement(n) => n.accept_metadata(visitor),
-        }
-    }
-
-    fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
-        visitor.visit_node_id(self.id())?;
-        Ok(())
-    }
-}
-
-generate_ast_methods!(
-    ArrayTypeName,
-    Assignment,
-    BinaryOperation,
-    Block,
-    Conditional,
-    ContractDefinition,
-    ElementaryTypeName,
-    ElementaryTypeNameExpression,
-    EmitStatement,
-    EnumDefinition,
-    EnumValue,
-    EventDefinition,
-    ErrorDefinition,
-    ExpressionStatement,
-    FunctionCall,
-    FunctionCallOptions,
-    FunctionDefinition,
-    FunctionTypeName,
-    ForStatement,
-    Identifier,
-    IdentifierPath,
-    IfStatement,
-    ImportDirective,
-    IndexAccess,
-    IndexRangeAccess,
-    InheritanceSpecifier,
-    InlineAssembly,
-    Literal,
-    MemberAccess,
-    NewExpression,
-    Mapping,
-    ModifierDefinition,
-    ModifierInvocation,
-    OverrideSpecifier,
-    ParameterList,
-    PragmaDirective,
-    Return,
-    RevertStatement,
-    SourceUnit,
-    StructDefinition,
-    StructuredDocumentation,
-    TryStatement,
-    TryCatchClause,
-    TupleExpression,
-    UnaryOperation,
-    UserDefinedTypeName,
-    UserDefinedValueTypeDefinition,
-    UsingForDirective,
-    VariableDeclaration,
-    VariableDeclarationStatement,
-    WhileStatement,
-    DoWhileStatement,
-    Break,
-    Continue,
-    PlaceholderStatement,
-);
-
-impl ASTNode {
-    pub fn src(&self) -> Option<&str> {
-        match self {
-            ASTNode::ArrayTypeName(node) => Some(&node.src),
-            ASTNode::Assignment(node) => Some(&node.src),
-            ASTNode::BinaryOperation(node) => Some(&node.src),
-            ASTNode::Block(node) => Some(&node.src),
-            ASTNode::Conditional(node) => Some(&node.src),
-            ASTNode::ContractDefinition(node) => Some(&node.src),
-            ASTNode::ElementaryTypeName(node) => Some(&node.src),
-            ASTNode::ElementaryTypeNameExpression(node) => Some(&node.src),
-            ASTNode::EmitStatement(node) => Some(&node.src),
-            ASTNode::EnumDefinition(node) => Some(&node.src),
-            ASTNode::EnumValue(node) => Some(&node.src),
-            ASTNode::EventDefinition(node) => Some(&node.src),
-            ASTNode::ErrorDefinition(node) => Some(&node.src),
-            ASTNode::ExpressionStatement(node) => Some(&node.src),
-            ASTNode::FunctionCall(node) => Some(&node.src),
-            ASTNode::FunctionCallOptions(node) => Some(&node.src),
-            ASTNode::FunctionDefinition(node) => Some(&node.src),
-            ASTNode::FunctionTypeName(node) => Some(&node.src),
-            ASTNode::ForStatement(node) => Some(&node.src),
-            ASTNode::Identifier(node) => Some(&node.src),
-            ASTNode::IdentifierPath(node) => Some(&node.src),
-            ASTNode::IfStatement(node) => Some(&node.src),
-            ASTNode::ImportDirective(node) => Some(&node.src),
-            ASTNode::IndexAccess(node) => Some(&node.src),
-            ASTNode::IndexRangeAccess(node) => Some(&node.src),
-            ASTNode::InheritanceSpecifier(node) => Some(&node.src),
-            ASTNode::InlineAssembly(node) => Some(&node.src),
-            ASTNode::Literal(node) => Some(&node.src),
-            ASTNode::MemberAccess(node) => Some(&node.src),
-            ASTNode::NewExpression(node) => Some(&node.src),
-            ASTNode::Mapping(node) => Some(&node.src),
-            ASTNode::ModifierDefinition(node) => Some(&node.src),
-            ASTNode::ModifierInvocation(node) => Some(&node.src),
-            ASTNode::OverrideSpecifier(node) => Some(&node.src),
-            ASTNode::ParameterList(node) => Some(&node.src),
-            ASTNode::PragmaDirective(node) => Some(&node.src),
-            ASTNode::Return(node) => Some(&node.src),
-            ASTNode::RevertStatement(node) => Some(&node.src),
-            ASTNode::SourceUnit(_) => None,
-            ASTNode::StructDefinition(node) => Some(&node.src),
-            ASTNode::StructuredDocumentation(node) => Some(&node.src),
-            ASTNode::TryStatement(node) => Some(&node.src),
-            ASTNode::TryCatchClause(node) => Some(&node.src),
-            ASTNode::TupleExpression(node) => Some(&node.src),
-            ASTNode::UnaryOperation(node) => Some(&node.src),
-            ASTNode::UserDefinedTypeName(node) => Some(&node.src),
-            ASTNode::UserDefinedValueTypeDefinition(node) => Some(&node.src),
-            ASTNode::UsingForDirective(node) => Some(&node.src),
-            ASTNode::VariableDeclaration(node) => Some(&node.src),
-            ASTNode::VariableDeclarationStatement(node) => Some(&node.src),
-            ASTNode::WhileStatement(node) => Some(&node.src),
-            ASTNode::DoWhileStatement(node) => Some(&node.src),
-            ASTNode::Break(node) => Some(&node.src),
-            ASTNode::Continue(node) => Some(&node.src),
-            ASTNode::PlaceholderStatement(node) => Some(&node.src),
-        }
-    }
-}
+use super::macros::generate_visit_methods_for_workspace_context_with_insert_node;
+pub use crate::ast::ASTNode;
 
 #[derive(Default, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct NodeContext {
@@ -529,182 +100,6 @@ impl WorkspaceContext {
 
     // Getters
 
-    pub fn array_type_names(&self) -> Vec<&ArrayTypeName> {
-        self.array_type_names_context.keys().collect()
-    }
-    pub fn assignments(&self) -> Vec<&Assignment> {
-        self.assignments_context.keys().collect()
-    }
-    pub fn binary_operations(&self) -> Vec<&BinaryOperation> {
-        self.binary_operations_context.keys().collect()
-    }
-    pub fn blocks(&self) -> Vec<&Block> {
-        self.blocks_context.keys().collect()
-    }
-    pub fn conditionals(&self) -> Vec<&Conditional> {
-        self.conditionals_context.keys().collect()
-    }
-    pub fn contract_definitions(&self) -> Vec<&ContractDefinition> {
-        self.contract_definitions_context.keys().collect()
-    }
-    pub fn elementary_type_names(&self) -> Vec<&ElementaryTypeName> {
-        self.elementary_type_names_context.keys().collect()
-    }
-    pub fn elementary_type_name_expressions(&self) -> Vec<&ElementaryTypeNameExpression> {
-        self.elementary_type_name_expressions_context
-            .keys()
-            .collect()
-    }
-    pub fn emit_statements(&self) -> Vec<&EmitStatement> {
-        self.emit_statements_context.keys().collect()
-    }
-    pub fn enum_definitions(&self) -> Vec<&EnumDefinition> {
-        self.enum_definitions_context.keys().collect()
-    }
-    pub fn enum_values(&self) -> Vec<&EnumValue> {
-        self.enum_values_context.keys().collect()
-    }
-    pub fn event_definitions(&self) -> Vec<&EventDefinition> {
-        self.event_definitions_context.keys().collect()
-    }
-    pub fn error_definitions(&self) -> Vec<&ErrorDefinition> {
-        self.error_definitions_context.keys().collect()
-    }
-    pub fn expression_statements(&self) -> Vec<&ExpressionStatement> {
-        self.expression_statements_context.keys().collect()
-    }
-    pub fn function_calls(&self) -> Vec<&FunctionCall> {
-        self.function_calls_context.keys().collect()
-    }
-    pub fn function_call_options(&self) -> Vec<&FunctionCallOptions> {
-        self.function_call_options_context.keys().collect()
-    }
-    pub fn function_definitions(&self) -> Vec<&FunctionDefinition> {
-        self.function_definitions_context.keys().collect()
-    }
-    pub fn function_type_names(&self) -> Vec<&FunctionTypeName> {
-        self.function_type_names_context.keys().collect()
-    }
-    pub fn for_statements(&self) -> Vec<&ForStatement> {
-        self.for_statements_context.keys().collect()
-    }
-    pub fn identifiers(&self) -> Vec<&Identifier> {
-        self.identifiers_context.keys().collect()
-    }
-    pub fn identifier_paths(&self) -> Vec<&IdentifierPath> {
-        self.identifier_paths_context.keys().collect()
-    }
-    pub fn if_statements(&self) -> Vec<&IfStatement> {
-        self.if_statements_context.keys().collect()
-    }
-    pub fn import_directives(&self) -> Vec<&ImportDirective> {
-        self.import_directives_context.keys().collect()
-    }
-    pub fn index_accesses(&self) -> Vec<&IndexAccess> {
-        self.index_accesses_context.keys().collect()
-    }
-    pub fn index_range_accesses(&self) -> Vec<&IndexRangeAccess> {
-        self.index_range_accesses_context.keys().collect()
-    }
-    pub fn inheritance_specifiers(&self) -> Vec<&InheritanceSpecifier> {
-        self.inheritance_specifiers_context.keys().collect()
-    }
-    pub fn inline_assemblies(&self) -> Vec<&InlineAssembly> {
-        self.inline_assemblies_context.keys().collect()
-    }
-    pub fn literals(&self) -> Vec<&Literal> {
-        self.literals_context.keys().collect()
-    }
-    pub fn member_accesses(&self) -> Vec<&MemberAccess> {
-        self.member_accesses_context.keys().collect()
-    }
-    pub fn new_expressions(&self) -> Vec<&NewExpression> {
-        self.new_expressions_context.keys().collect()
-    }
-    pub fn mappings(&self) -> Vec<&Mapping> {
-        self.mappings_context.keys().collect()
-    }
-    pub fn modifier_definitions(&self) -> Vec<&ModifierDefinition> {
-        self.modifier_definitions_context.keys().collect()
-    }
-    pub fn modifier_invocations(&self) -> Vec<&ModifierInvocation> {
-        self.modifier_invocations_context.keys().collect()
-    }
-    pub fn override_specifiers(&self) -> Vec<&OverrideSpecifier> {
-        self.override_specifiers_context.keys().collect()
-    }
-    pub fn parameter_lists(&self) -> Vec<&ParameterList> {
-        self.parameter_lists_context.keys().collect()
-    }
-    pub fn pragma_directives(&self) -> Vec<&PragmaDirective> {
-        self.pragma_directives_context.keys().collect()
-    }
-    pub fn returns(&self) -> Vec<&Return> {
-        self.returns_context.keys().collect()
-    }
-    pub fn revert_statements(&self) -> Vec<&RevertStatement> {
-        self.revert_statements_context.keys().collect()
-    }
-    pub fn source_units(&self) -> Vec<&SourceUnit> {
-        self.source_units_context.iter().collect()
-    }
-    pub fn struct_definitions(&self) -> Vec<&StructDefinition> {
-        self.struct_definitions_context.keys().collect()
-    }
-    pub fn structured_documentations(&self) -> Vec<&StructuredDocumentation> {
-        self.structured_documentations_context.keys().collect()
-    }
-    pub fn try_statements(&self) -> Vec<&TryStatement> {
-        self.try_statements_context.keys().collect()
-    }
-    pub fn try_catch_clauses(&self) -> Vec<&TryCatchClause> {
-        self.try_catch_clauses_context.keys().collect()
-    }
-    pub fn tuple_expressions(&self) -> Vec<&TupleExpression> {
-        self.tuple_expressions_context.keys().collect()
-    }
-    pub fn unary_operations(&self) -> Vec<&UnaryOperation> {
-        self.unary_operations_context.keys().collect()
-    }
-    pub fn user_defined_type_names(&self) -> Vec<&UserDefinedTypeName> {
-        self.user_defined_type_names_context.keys().collect()
-    }
-    pub fn user_defined_value_type_definitions(&self) -> Vec<&UserDefinedValueTypeDefinition> {
-        self.user_defined_value_type_definitions_context
-            .keys()
-            .collect()
-    }
-    pub fn using_for_directives(&self) -> Vec<&UsingForDirective> {
-        self.using_for_directives_context.keys().collect()
-    }
-    pub fn variable_declarations(&self) -> Vec<&VariableDeclaration> {
-        self.variable_declarations_context.keys().collect()
-    }
-    pub fn variable_declaration_statements(&self) -> Vec<&VariableDeclarationStatement> {
-        self.variable_declaration_statements_context
-            .keys()
-            .collect()
-    }
-    pub fn while_statements(&self) -> Vec<&WhileStatement> {
-        self.while_statements_context.keys().collect()
-    }
-
-    pub fn do_while_statements(&self) -> Vec<&DoWhileStatement> {
-        self.do_while_statements_context.keys().collect()
-    }
-
-    pub fn break_statements(&self) -> Vec<&Break> {
-        self.break_statements_context.keys().collect()
-    }
-
-    pub fn continue_statements(&self) -> Vec<&Continue> {
-        self.continue_statements_context.keys().collect()
-    }
-
-    pub fn placeholder_statements(&self) -> Vec<&PlaceholderStatement> {
-        self.placeholder_statements_context.keys().collect()
-    }
-
     pub fn get_parent(&self, node_id: NodeID) -> Option<&ASTNode> {
         self.nodes.get(self.parent_link.get(&node_id)?)
     }
@@ -769,236 +164,6 @@ impl WorkspaceContext {
             return Some((offset, len));
         }
         None
-    }
-
-    pub fn get_source_unit_from_child_node(&self, node: &ASTNode) -> Option<&SourceUnit> {
-        let source_unit_id = match node {
-            ASTNode::ArrayTypeName(node) => self
-                .array_type_names_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Assignment(node) => self
-                .assignments_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::BinaryOperation(node) => self
-                .binary_operations_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Block(node) => self
-                .blocks_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Conditional(node) => self
-                .conditionals_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ContractDefinition(node) => self
-                .contract_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ElementaryTypeName(node) => self
-                .elementary_type_names_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ElementaryTypeNameExpression(node) => self
-                .elementary_type_name_expressions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::EmitStatement(node) => self
-                .emit_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::EnumDefinition(node) => self
-                .enum_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::EnumValue(node) => self
-                .enum_values_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::EventDefinition(node) => self
-                .event_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ErrorDefinition(node) => self
-                .error_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ExpressionStatement(node) => self
-                .expression_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::FunctionCall(node) => self
-                .function_calls_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::FunctionCallOptions(node) => self
-                .function_call_options_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::FunctionDefinition(node) => self
-                .function_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::FunctionTypeName(node) => self
-                .function_type_names_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ForStatement(node) => self
-                .for_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Identifier(node) => self
-                .identifiers_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::IdentifierPath(node) => self
-                .identifier_paths_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::IfStatement(node) => self
-                .if_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ImportDirective(node) => self
-                .import_directives_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::IndexAccess(node) => self
-                .index_accesses_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::IndexRangeAccess(node) => self
-                .index_range_accesses_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::InheritanceSpecifier(node) => self
-                .inheritance_specifiers_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::InlineAssembly(node) => self
-                .inline_assemblies_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Literal(node) => self
-                .literals_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::MemberAccess(node) => self
-                .member_accesses_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::NewExpression(node) => self
-                .new_expressions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Mapping(node) => self
-                .mappings_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ModifierDefinition(node) => self
-                .modifier_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ModifierInvocation(node) => self
-                .modifier_invocations_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::OverrideSpecifier(node) => self
-                .override_specifiers_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::ParameterList(node) => self
-                .parameter_lists_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::PragmaDirective(node) => self
-                .pragma_directives_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Return(node) => self
-                .returns_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::RevertStatement(node) => self
-                .revert_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::SourceUnit(node) => Some(node.id),
-            ASTNode::StructDefinition(node) => self
-                .struct_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::StructuredDocumentation(node) => self
-                .structured_documentations_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::TryStatement(node) => self
-                .try_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::TryCatchClause(node) => self
-                .try_catch_clauses_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::TupleExpression(node) => self
-                .tuple_expressions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::UnaryOperation(node) => self
-                .unary_operations_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::UserDefinedTypeName(node) => self
-                .user_defined_type_names_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::UserDefinedValueTypeDefinition(node) => self
-                .user_defined_value_type_definitions_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::UsingForDirective(node) => self
-                .using_for_directives_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::VariableDeclaration(node) => self
-                .variable_declarations_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::VariableDeclarationStatement(node) => self
-                .variable_declaration_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::WhileStatement(node) => self
-                .while_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::DoWhileStatement(node) => self
-                .do_while_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Break(node) => self
-                .break_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::Continue(node) => self
-                .continue_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-            ASTNode::PlaceholderStatement(node) => self
-                .placeholder_statements_context
-                .get(node)
-                .map(|context| context.source_unit_id),
-        };
-
-        // iterate through self.source_units until the source unit with the id matching `source_unit_id` is found, then return its `absolute_path`
-
-        source_unit_id.and_then(|id| {
-            self.source_units_context
-                .iter()
-                .find(|source_unit| source_unit.id == id)
-        })
     }
 
     pub fn get_node_sort_key_from_capturable(
